Build cuda support
Since cuda support is built with the conda package of hoomd, this also
means that the package needs to be built with that support. This is an
issue since the objects in the include files have conditionally compiled
components, which would have resulted in a mismatch in the structure of
the resulting classes when trying to pass objects between each binary.